in condition

+++ src/modules/bloggers-platform/application/usecases/blogs/delete-blog.usecase.ts
@@ -1,8 +1,13 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
+import { DomainExceptionCode } from 'src/core/exceptions/domain-exception-codes';
+import { DomainException } from 'src/core/exceptions/domain-exceptions';
 import { BlogsRepository } from 'src/modules/bloggers-platform/infrastructure/blogs.repository';
 
 export class DeleteBlogCommand {
-  constructor(public id: string) {}
+  constructor(
+    public id: string,
+    public userId?: string,
+  ) {}
 }
 
 @CommandHandler(DeleteBlogCommand)
@@ -11,7 +16,16 @@ export class DeleteBlogUseCase
 {
   constructor(private blogsRepository: BlogsRepository) {}
 
-  async execute({ id }: DeleteBlogCommand): Promise<void> {
-    await this.blogsRepository.deleteOrNotFoundFail(id);
+  async execute({ id, userId }: DeleteBlogCommand): Promise<void> {
+    const blog = await this.blogsRepository.findOrNotFoundFail(id);
+
+    if (blog.userId === Number(userId)) {
+      await this.blogsRepository.deleteOrNotFoundFail(id);
+    } else {
+      throw new DomainException({
+        code: DomainExceptionCode.Forbidden,
+        message: 'Blog was created by another user',
+      });
+    }
   }
 }
